requirements.txt
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.17.0
pandas>=2.2.2
pandas-gbq>=0.22.0
scikit-learn>=1.5.1
matplotlib>=3.8.4